Move WallCellNode to separate file
